Add files via upload
Uncommented final fast algorithm. The txt file is what you should get when running it for the 3x3x4 case, with reference to the numbering in my thesis. It should only take a minute or two to run this file on your computer for the 3x3x4 case, it scales reasonably well for larger cases but will still be noticeably slower.